@@ -22,12 +22,14 @@
     "jsonwebtoken": "^9.0.2",
     "mongodb": "^6.5.0",
     "mongoose": "^8.3.1",
+    "mongoose-paginate-v2": "^1.8.3",
     "nodemailer": "^6.9.14",
     "nodemon": "^3.1.0",
     "passport": "^0.7.0",
     "passport-github2": "^0.1.12",
     "passport-local": "^1.0.0",
     "socket.io": "^4.7.5",
+    "stripe": "^16.8.0",
     "swagger-jsdoc": "^6.2.8",
     "swagger-ui-express": "^5.0.1",
     "winston": "^3.13.0"
@@ -35,7 +37,7 @@
   "devDependencies": {
     "@eslint/js": "^9.5.0",
     "chai": "^5.1.1",
-    "eslint": "^9.5.0",
+    "eslint": "^9.9.0",
     "esm": "^3.2.25",
     "globals": "^15.6.0",
     "jest": "^29.7.0",
