@@ -3,7 +3,7 @@ export default [
     ignores: ['node_modules/**'],
   },
   {
-    files: ['**/*.js'],
+    files: ['**/*.js', '**/*.jsx', '**/*.ts', '**/*.tsx'],
     languageOptions: {
       globals: {
         require: 'readonly',
@@ -12,10 +12,17 @@ export default [
         process: 'readonly',
         __dirname: 'readonly',
       },
+      ecmaVersion: 2021,
+      sourceType: 'module',
     },
     rules: {
-      'no-unused-vars': 'warn',
+      'no-unused-vars': ['warn', { "vars": "all", "args": "after-used", "ignoreRestSiblings": false }],
       'no-console': 'off',
+      'no-debugger': 'warn',
+      'no-var': 'error',
+      'prefer-const': 'error',
+      'eqeqeq': 'error',
+      'no-duplicate-imports': 'error',
     },
   },
 ];
