@@ -1,8 +1,21 @@
-const isUser = (req, res, next) => {
-  if (req.isAuthenticated() && req.user.role === 'user') {
-    return next();
+import User from '../../models/user.js';
+
+const isUser = async (req, res, next) => {
+  if (!req.isAuthenticated || !req.isAuthenticated()) {
+    return res.status(401).json({ message: 'No estás autenticado' });
   }
-  res.status(403).json({ message: 'Access denied. User role required.' });
+
+  try {
+    const user = await User.findById(req.user._id);
+    if (!user) {
+      return res.status(404).json({ message: 'Usuario no encontrado' });
+    }
+
+    req.user = user;
+    next();
+  } catch {
+    res.status(500).json({ message: 'Error del servidor' });
+  }
 };
 
 export default isUser;
