@@ -1,6 +1,6 @@
 const validateStringFields = (fields) => {
   return (req, res, next) => {
-    for (let field of fields) {
+    for (const field of fields) {
       if (typeof req.body[field] !== 'string' || req.body[field].trim() === '') {
         return res.status(400).json({ message: `Invalid ${field}` });
       }
