@@ -1,6 +1,7 @@
 const express = require("express");
 const router = express.Router();
 const Message = require("./message");
+import logger from '/config/logger';
 
 // Endpoint para recibir y guardar mensajes
 router.post("/messages", async (req, res) => {
@@ -10,7 +11,7 @@ router.post("/messages", async (req, res) => {
     await newMessage.save();
     res.status(201).json({ message: "Message sent successfully" });
   } catch (error) {
-    console.error("Error sending message:", error);
+    logger.error("Error sending message:", error);
     res.status(500).json({ error: "Error sending message" });
   }
 });
