@@ -1,43 +1,19 @@
 import mongoose from 'mongoose';
+import mongoosePaginate from 'mongoose-paginate-v2';
 
 const productSchema = new mongoose.Schema({
-  title: {
-    type: String,
-    required: true
-  },
-  description: {
-    type: String,
-    required: true
-  },
-  code: {
-    type: String,
-    required: true,
-    unique: true
-  },
-  price: {
-    type: Number,
-    required: true
-  },
-  status: {
-    type: Boolean,
-    required: true,
-    default: true
-  },
-  stock: {
-    type: Number,
-    required: true
-  },
-  category: {
-    type: String,
-    required: true
-  },
-  thumbnails: [
-    {
-      type: String,
-      required: true
-    }
-  ]
+  title: { type: String, required: true },
+  description: { type: String, required: true },
+  code: { type: String, required: true },
+  price: { type: Number, required: true },
+  stock: { type: Number, required: true },
+  category: { type: String, required: true },
+  thumbnails: [String],
+  status: { type: Boolean, default: true },
+  owner: { type: String, default: 'admin' },
 }, { timestamps: true });
+
+productSchema.plugin(mongoosePaginate);
 
 const Product = mongoose.model('Product', productSchema);
 
